move code to related block
